requirements.txt
# Cloud Functions Framework
functions-framework==3.*

# xAI SDK for Agent Tools API (Grok sentiment analysis)
xai-sdk>=1.5.0

# Time/Date handling
pytz>=2024.1

# HTTP requests (for Discord webhook)
requests>=2.31.0

# Data analysis and manipulation
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud Storage (for reading historical logs)
google-cloud-storage>=2.10.0

# Environment variables
python-dotenv>=1.0.0
